requirements.txt
openai
chromadb
numpy
pyyaml
requests
beautifulsoup4
google-search-results
